Moved the prompts into a utils
